Add float mutation and test for float GA

+++ individual.h
@@ -41,7 +41,7 @@ public:
 
 private:
     static constexpr size_t chromosomes = (genome_size + 63)/64;
-    static constexpr uint64_t final_chromosome_mask = genome_size % 64 == 0 ? ~0 : ~0 << 64 - ( genome_size % 64 );
+    static constexpr uint64_t final_chromosome_mask = genome_size % 64 == 0 ? ~0 : (~((uint64_t)0)) << 64 - ( genome_size % 64 );
     double mut_step = 1;
 
     std::array<Chromosome, chromosomes> genes;
